Locked ball to paddle and can now launch it with a left click
The ball remains attached to the paddle and only launches with a left click. (accessing the ball's Rigidbody component and giving it a velocity to thrust upwards). To lock the ball to the paddle, we get the distance between the paddle and ball. Then we retrieve the current x and y positions of the paddle, which we will use in conjuction with the distance to set the position of the ball.